Update data (with wc switch), fix active cases
Western Cape data was completed backdated to the earliest data that
could be found.  The data source for the Western Cape was also switched
from the daily data from the NICD to the data communicated daily by the
WC government, since they give information on the number of tests
performed in the province.  It should be noted that this data is
released daily, giving the statistics as at 1pm.  This is recorded here
as the under the previous day's date, under the assumption that most
data taken up between midnight and 1pm would just be information of the
previous day uploaded after the fact.

Active cases was until now wrongly calculated as the difference between
positive cases and recovery, where the correct calculation should also
subtract the number of deaths.

Another small fix is where lines in the data is commented out with '#':
until now this wrongfully ended data parsing after the commented line,
whereas it now simply skips the line, as it should have done from the
beginning.